Adding online documentation deployment
Oneline documentation deployment to gh-pages has been added by adding
deploydocs function to the make.jl script. The deployment only happens
when the script is called from inside the GitHub Actions workflow.

The workflow is defined in the .github/workflows/make_docs.yml file. It
initializes the separate environment for the documentation in the docs
directory and adds the current project version in develop mode. Thus,
the ./docs/Project.toml should not have IMASDD hash otherwise the docs
will be generated for a fixed version always.

I also removed the docs/src/api.md file as it is an auto-generated file
by make.jl and thus should be ignored by git.